feat: merge one file
