+++ src/time_travel.py
@@ -1,110 +1,123 @@
-"""
-要件
-
-タスク発注するパターンを3つに分ける
-ファイルリストを作成する
-ファイルリストを使って、タスク発注する
-
-タスク発注は、datetimeを使って現時刻を取得し、タスク発注する時間になったら、
-現時刻から指定時間前の時間から現時刻までの間にを発注する
-
-"""
-
-import configparser
-import ast
 import datetime
-import random
+import pathlib
+import yaml
 
-def get_service_number() -> int:
-    """サービス番号を取得する
+# task.ymlのパスを取得: pathlibをつかって
+# config は、src と同じ階層にあるとする
+task_path = pathlib.Path(__file__).parent.cwd() / 'config' / 'task.yml'
 
-    Returns:
-        [int]: サービス番号を取得する
-
-    """
-    service_number = random.randint(1, 15)
-    return service_number
-
-def read_file(file_name: str, section: str, tasks_key: str, service_numbers_key=None) -> tuple:
-    """ファイルを読み込む
+def read_file(read_file) -> dict:
+    """yaml ファイルを読み込んで、dict に変換して返す
 
     Args:
-        file_name (str): ファイル名
-        section (str): セクション名
-        tasks_key (str): タスクキー
-        service_numbers_key (str, optional): サービス番号キー. Defaults to None.
+        read_file (str): 読み込むファイルのパス
 
     Returns:
-        [tuple]: タスクとサービス番号を返す
-
+        dict: yaml ファイルの中身
     """
+    with open(read_file, 'r', encoding='utf-8') as f:
+        return yaml.safe_load(f)
 
-    config_ini = configparser.ConfigParser()
-    config_ini.read(file_name, encoding='utf-8')
-
-    tasks_str = config_ini[section][tasks_key]
-    tasks = ast.literal_eval(tasks_str)
-
-    # もし service_numbers_key がある場合
-    if service_numbers_key is not None:
-        service_numbers_str = config_ini[section][service_numbers_key]
-        service_numbers = ast.literal_eval(service_numbers_str)
-        return tasks, service_numbers
-    else:
-        return tasks, None
-
-def target_task_time(now: datetime, service_number: int) -> int:
-    """発注する時間を取得する
+def process_patterns(patterns) -> dict:
+    """patterns を処理して、settings を返す
 
     Args:
-        now (datetime): 現在時刻
-        service_number (int): サービス番号
+        patterns (dict): patterns の中身
 
     Returns:
-        [int]: 発注する時間を取得する
+        dict: settings の中身
+    """
+    settings = {}
 
+    for pattern_name, pattern_data in patterns.items():
+        intervals = pattern_data.get('interval')
+        service_numbers = pattern_data.get('service_numbers')
+
+        # コメント に記載のある json 形式で settings に追加
+        settings[pattern_name] = {
+            'interval': intervals,
+            'service_numbers': service_numbers
+        }
+
+    return settings
+
+def target_service_interval(service_number, settings):
+    """service_number に対応する interval を返す
+
+    Args:
+        service_number (str): サービス番号
+        settings (dict): settings の中身
+
+    Returns:
+        dict: interval の中身
     """
 
-    tasks_first, service_numbers_first = read_file('tasks.ini', 'FIRST_TASKS', 'tasks', 'service_numbers')
-    tasks_second, service_numbers_second = read_file('tasks.ini', 'SECOND_TASKS', 'tasks', 'service_numbers')
-    tasks_default, _ = read_file('tasks.ini', 'DEFAULT', 'tasks', None)
+    sections = [section for section in settings]
+    task_section = 'default'
 
-    print({'service_number:': service_number})
-    if service_number in service_numbers_first:
-        # pattern1
-        time_back = time_tasks(now, tasks_first)
-        if time_back:
-            return time_back + 10
-        return 0
-    elif service_number in service_numbers_second:
-        # pattern2
-        time_back = time_tasks(now, tasks_second)
-        if time_back:
-            return time_back + 10
-        return 0
-    else:
-        # default
-        time_back = time_tasks(now, tasks_default)
-        if time_back:
-            return time_back + 10
-        return 0
+    for section in sections:
+        if settings[section]['service_numbers'] is not None\
+            and service_number in settings[section]['service_numbers']:
+            task_section = section
+            break
+    return {
+        'service_number': service_number,
+        'section': task_section,
+        'interval': settings[task_section]['interval']
+        }
 
-def result_time(now: datetime) -> tuple[int, int]:
-    """発注する時間を取得する
+def _get_time_interval(now, tasks) -> int:
+    """タスク発注する時間を取得する
 
     Args:
         now (datetime): 現在時刻
+        tasks (list): タスク発注する時間のリスト
 
     Returns:
-        [tuple[int, int]]: 発注する時間を取得する
+        [int]: タスク発注する時間を取得する
 
     """
-    target_time = target_task_time(now, service_number=get_service_number())
-    if target_time == 0:
-        return 0, 0
- 
-    # 発注する時間を取得
-    since = datetime.datetime.combine(datetime.date.today(), now.time()) - datetime.timedelta(minutes=target_time)
+    # Define the mocked time interval
+    time_interval: int = 0
+
+    hour, minute = now.strftime("%H,%M").split(",")
+    minute = int(int(minute) / 10) * 10
+    # tasks に格納されている値に, 10分単位で丸める.
+    # 意図は、実行時間を10分単位で丸めることで、実行時間の猶予を9min作る.
+    task_time = int(hour) + minute / 60
+
+    if not task_time in tasks:
+        return time_interval
+
+    """NOTE: コード理解のため.分割して書いた """
+    # taeget task index
+    task_index = tasks.index(task_time)
+    # before task time diffrence
+    before_time_diff = tasks[task_index] - tasks[(task_index -1)%len(tasks)]
+    # convert to hours
+    _ = int(before_time_diff%24 * 60)
+
+    # one line
+    time_interval = int(((tasks[tasks.index(task_time)] - tasks[(tasks.index(task_time) -1)%len(tasks)])%24) *60)
+
+    return time_interval
+
+def get_task_time_interval(now: datetime, time_interval) -> tuple:
+    """タスク発注する時間を取得する
+
+    Args:
+        now (datetime): 現在時刻
+        time_interval (int): タスク発注する時間のリスト
+
+    Returns:
+        [int]: タスク発注する時間を取得する. 時間外の場合は、None を返す.
+
+    """
+
+    if time_interval == 0:
+        return None, None
+
+    since = datetime.datetime.combine(datetime.date.today(), now.time()) - datetime.timedelta(minutes=time_interval)
 
     # str型に変換
     since = since.strftime('%Y-%m-%dT%H:%M:%S')
@@ -113,10 +126,19 @@ def result_time(now: datetime) -> tuple[int, int]:
     return since, until
 
 def main():
-    """メイン処理"""
-    now = datetime.datetime.now()
-    since, until = result_time(now)
-    print({'since:': since, 'until:': until})
+    """main function"""
+    # now = datetime.now()
+    mock_now = datetime.datetime(2021, 4, 19, 0, 8, 0)
+    task = read_file(task_path)
+    settings = process_patterns(task['time_interval'])
+    service_numbers = ['b001', 'b005', 'b008']
+
+    for service_number in service_numbers:
+        setting = target_service_interval(service_number, settings)
+        time_interval = _get_time_interval(mock_now, setting['interval'])
+        since, until = get_task_time_interval(mock_now, time_interval)
+
+        print(f'service_number: {service_number}, section: {setting["section"]}, since: {since}, until: {until}')
 
 
 if __name__ == '__main__':
